Adding attributes for locations of x at label. Treating label in cntex.
During typing, this adds attributes for all variable under an at/old. The
attributes are then propagated inside the ident of the constants
corresponding to them during a new transformation (only in
counterexamples phase) attr_to_ls. So the counterexamples are printed with
new attributes that corresponds to the labels they had.
We use this information to correctly print "x'old" and "x at L" in the
counterexamples final output.

Also adds new example bench/ce/records_label.mlw
Bench not updated